Added test cases in chat_service_test and channel_client_test for the coverage issue.
